User - 마이페이지 -> 관심상품, 내 게시글보기 완료.

+++ src/main/java/hello/aiofirstuser/service/WishProductServiceImpl.java
@@ -3,6 +3,7 @@ package hello.aiofirstuser.service;
 import hello.aiofirstuser.domain.Member;
 import hello.aiofirstuser.domain.ProductVariant;
 import hello.aiofirstuser.domain.WishProduct;
+import hello.aiofirstuser.dto.product.WishProductRequestListDTO;
 import hello.aiofirstuser.dto.product.WishProductResponseDTO;
 import hello.aiofirstuser.repository.MemberRepository;
 import hello.aiofirstuser.repository.ProductVariantRepository;
@@ -10,7 +11,9 @@ import hello.aiofirstuser.repository.WishProductRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -24,6 +27,7 @@ public class WishProductServiceImpl implements WishProductService {
     private final MemberRepository memberRepository;
 
     @Override
+    @Transactional
     public int save(Long productVariantId, Member member) {
         Optional<ProductVariant> result = productVariantRepository.findById(productVariantId);
 
@@ -43,19 +47,58 @@ public class WishProductServiceImpl implements WishProductService {
     }
 
     @Override
-    public void remove(Long wishProductId, Long memberId) {
+    @Transactional
+    public int remove(Long wishProductId, Member member) {
 
-        Optional<WishProduct> result = wishProductRepository.findByMemberIdAndId(wishProductId, memberId);
+        Optional<WishProduct> result = wishProductRepository.findByMemberIdAndId(wishProductId, member.getId());
 
-        WishProduct wishProduct = result.orElseThrow();
+        WishProduct wishProduct = result.orElse(null);
 
-        wishProductRepository.delete(wishProduct);
+        if(wishProduct != null){
+            wishProductRepository.delete(wishProduct);
+            return 1;
+        }
+
+        return 0;
 
     }
 
     @Override
-    public List<WishProductResponseDTO> getMemberWishList(Long memberId) {
+    @Transactional
+    public int removeAll(WishProductRequestListDTO wishProductRequestListDTO, Member member) {
+        List<Long> wishIds = wishProductRequestListDTO.getWishProductRequestDTOS()
+                .stream().map(wishProductRequestDTO -> wishProductRequestDTO.getWishProductId()).toList();
 
-        return null;
+        List<WishProduct> wishProducts = wishProductRepository.findByIdsAndMemberId(wishIds, member.getId());
+        int count = 0;
+
+        if (!wishProducts.isEmpty()){
+            for (WishProduct wishProduct : wishProducts){
+                wishProductRepository.delete(wishProduct);
+                count++;
+            }
+
+            if (wishProducts.size() == count ){
+                return 1;
+            }
+        }
+
+        return 0;
+    }
+
+    @Override
+    public List<WishProductResponseDTO> getMemberWishList(Member member) {
+
+        List<WishProduct> wishProducts = wishProductRepository.findByMemberId(member.getId());
+
+        List<WishProductResponseDTO> wishProductResponseDTOS = new ArrayList<>();
+
+        if (!wishProducts.isEmpty()){
+            for (WishProduct wishProduct : wishProducts){
+                wishProductResponseDTOS.add(entityToWishProductResponseDTO(wishProduct));
+            }
+        }
+
+        return wishProductResponseDTOS;
     }
 }
